Fixed bug in midi_length function in OS X implementation.

+++ portmidi/trunk/porttime/ptwinmm.c
@@ -64,7 +64,7 @@ PtTimestamp Pt_Time()
 }
 
 
-void Pt_Sleep(int duration)
+void Pt_Sleep(long duration)
 {
     Sleep(duration);
 }
